Fix compiler warning caused by type mismatch between kore and kinc.

+++ Sources/Kore/Math/Random.cpp
@@ -4,18 +4,18 @@
 
 using namespace Kore;
 
-void Random::init(int seed) {
+void Random::init(int64_t seed) {
 	kinc_random_init(seed);
 }
 
-s32 Random::get() {
+int64_t Random::get() {
 	return kinc_random_get();
 }
 
-s32 Random::get(s32 max) {
+int64_t Random::get(int64_t max) {
 	return kinc_random_get_max(max);
 }
 
-s32 Random::get(s32 min, s32 max) {
+int64_t Random::get(int64_t min, int64_t max) {
 	return kinc_random_get_in(min, max);
 }
